backend/app/requirements.txt
Flask==2.3.2
redis==4.6.0
python-dotenv==1.0.0
psycopg2==2.9.7